tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/003.html

<!--
var x = 'a';
try {
  importScripts('data:text/javascript,x+="b"',
                'data:text/javascript,x+="c"');
} catch(e) {
  x += "d"
}
postMessage(x);
/*
-->
<!doctype html>
<title>importScripts running scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, "abc");
    this.done();
  });
});
</script>
<!--
*/
//-->
